[FIx] : storage 데이터 삭제 문제 수정

+++ Poki-iOS/Poki-iOS/Sources/Presenter/MyPage/Setting/AccountDeletionVC.swift
@@ -210,6 +210,8 @@ final class AccountDeletionVC: UIViewController {
     }
     
     func deleteAllDatas() {
+        guard let imageUrl = firestoreManager.userData?.imageURL else { return }
+        StorageManager.shared.deleteImage(imageURL: imageUrl) { _ in }
         firestoreManager.deleteAllPhotoData()
         firestoreManager.deleteAllPoseData()
         firestoreManager.deleteUserDocument()
